moving handling progress detection to ProgressSlider, changing notifications to use float

+++ gui/com/frostwire/gui/mplayer/ProgressSliderListener.java
@@ -2,7 +2,7 @@ package com.frostwire.gui.mplayer;
 
 public interface ProgressSliderListener {
 
-	public void onProgressSliderTimeValueChange( int seconds );
+	public void onProgressSliderTimeValueChange( float seconds );
 	public void onProgressSliderMouseDown();
 	public void onProgressSliderMouseUp();
 }
